Added alternate versions

+++ app/src/routes/deleteItem.js
@@ -1,6 +1,10 @@
 const db = require('../persistence');
 
 module.exports = async (req, res) => {
-    await db.removeItem(req.params.id);
+    if(!process.env.READ)
+    {
+        await db.removeItem(req.params.id);
+    }
+    
     res.sendStatus(200);
 };
